% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{broken_stick}
\alias{broken_stick}
\title{Broken stick regression}
\usage{
broken_stick(response, predictor, split.point)
}
\arguments{
\item{response}{The response variable (a vector)}

\item{predictor}{The predictor variable (a vector)}

\item{split.point}{The point to split the broken stick regression}
}
\description{
One class broken stick regression
}
\examples{
broken_stick(response, predictor, 50)
}
\keyword{broken}
\keyword{stick}

