% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/finance_tools.R
\name{calculate_dcgr}
\alias{calculate_dcgr}
\title{A R function that calculates the discrete continuous growth rate of a stock}
\usage{
calculate_dcgr(ticker, from = (Sys.Date() - 365), to = Sys.Date())
}
\arguments{
\item{ticker}{The stock ticker for a given company}

\item{from}{The start date for the stock data (year month date format)}

\item{to}{The end date for the stock date (year month date format)}
}
\description{
This function allows you to specify a stock ticker and date ranges such that the data is grabbed from the
yahoo finance API. The output consists of the mean dcgr and the a vector of the logged differences
}
\examples{
dcgr <- calculate_dcgr("goog")
}
\keyword{dcgr}
\keyword{finance,}

