% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/finance_tools.R
\name{calculate_idv}
\alias{calculate_idv}
\title{A R function that calls calculates the Implied Daily Volatility (IDV) of a stock}
\usage{
calculate_idv(strike, stock_price, interest, time, option_price,
  type = "call")
}
\arguments{
\item{strike}{The strike price of astock}

\item{stock_price}{The stock price of the stock}

\item{interest}{The interest rate}

\item{time}{The time until execution of the option}

\item{option_price}{The price of the option}
}
\description{
This function calculates the IDV using the Black-Scholes method. The user must specify the strike
price, the stock price, the interest rate, time until execution, option price (call or put), and the type
of option (call or put)
}
\examples{
call_idv <- calculate_idv(113, 111.5, 0.02, 17, 3.50, "call")
}
\keyword{finance,}
\keyword{idv}

