% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_diag_tools.R
\name{error_check}
\alias{error_check}
\title{Diagnostics of errors}
\usage{
error_check(model)
}
\arguments{
\item{model}{A linear model to diagnose, of class 'lm'}
}
\description{
Checks the errors/residuals with the Durbin Watts statistic (serial correlation), variation inflation factors and condition numbers
}
\examples{
error_check(lm(resp ~ pred, data))
}
\keyword{diagnostics}
\keyword{errors,}

