% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{lm_ref}
\alias{lm_ref}
\title{A lookup for various regression methods}
\usage{
lm_ref()
}
\description{
This function allows you to lookup which package/method corresponds to a given regression method. Returns a data frame
 with the package and method that each method matches up with, as well as a short description of each method.
}
\examples{
lm_ref()
}
\keyword{reference}

