% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{my_pcr}
\alias{my_pcr}
\title{A wrapper for principal component regression}
\usage{
my_pcr(lin_mod)
}
\arguments{
\item{lin_mod}{A linear model of class 'lm'}
}
\description{
Does PCR the easy way
}
\examples{
my_pcr(lm(resp ~ predictor))
}
\keyword{pcr}

