% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{my_pls}
\alias{my_pls}
\title{A wrapper for partial least squares regression.}
\usage{
my_pls(lin_mod, ncomp = 5)
}
\arguments{
\item{lin_mod}{A linear model of class 'lm'}

\item{ncomp}{The number of components to consider in the PLS regression}
}
\description{
Does partial least squares.
}
\examples{
my_pls(lm(resp ~ predictor), ncomp =2)
}
\keyword{pls}

