% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_diag_tools.R
\name{normality_check}
\alias{normality_check}
\title{Linear model normality diagnostic}
\usage{
normality_check(model)
}
\arguments{
\item{model}{A linear model of class 'lm'}
}
\description{
Computes model diagnostics for a linear model, including shapiro-wilkes,
influential observations, outliers, and large leverage points
}
\examples{
normality_check(lm(resp ~ predictor))
}
\keyword{diagnostics}
\keyword{normality,}

