% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/finance_tools.R
\name{portfolio_analysis}
\alias{portfolio_analysis}
\title{A R function that calculates a bunch of relevant portfolio statistics, given an array of stock tickers}
\usage{
portfolio_analysis(tickers, ratios = rep(1/length(tickers), length(tickers)))
}
\arguments{
\item{tickers}{A vector of stock tickers (the portfolio).}

\item{ratios}{The ratios of the stocks in the portfolio. A vcetor}
}
\description{
This function calculates portfolio variance & volatility, portfolio and annualized alpha,
unweighted variance/volatility, and the correlation matrix.
}
\examples{
my_portfolio <- portfolio_analysis(tickers=c("SPY", "AAPL", "GOOG", "GLD"),
                                   ratios=c(0.5,0.2,0.2,0.1))
}
\keyword{finance,}
\keyword{portfolio,}
\keyword{variance}

