% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/lm_tools.R
\name{stepwise}
\alias{stepwise}
\title{Stepwise Regression wrapper}
\usage{
stepwise(mod, type = "AIC")
}
\arguments{
\item{mod}{The linear model of class 'lm'}

\item{type}{The type of stepwise regression. Default is AIC but other options include "Adj-R" for adjusted
r-squared, and "Mallows" for mallows critical point}
}
\description{
This function allows you to perform stepwise regression in three different ways
}
\examples{
stepwise(lm(resp~pred, data), type = "Adj-R")
}
\keyword{stepwise}

