% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/finance_tools.R
\name{yahoo_finance}
\alias{yahoo_finance}
\title{A R function that calls the Yahoo Finance API to get data into a data frame}
\usage{
yahoo_finance(ticker, date.from = "2015-01-01", data.to = Sys.Date())
}
\arguments{
\item{ticker}{The stock ticker for a given company}

\item{date.from}{The start date for the stock data (year month date format)}

\item{date.to}{The end date for the stock date (year month date format)}
}
\description{
This function allows you to specify a stock ticker and date ranges such that the data is grabbed from the
yahoo finance API. The output is a data.frame of all of the information returned.
}
\examples{
data <- yahoo_finance("goog")
plot.ts(data$Adj.Close)
}
\keyword{finance}

